Added 'auth_avatar' table migration. Added 'AvatarB64' schema for POST-Endpoint Body. Updated 'upload_user_avatar' according to new schema and added bool return for 'delete_avatar'. Added case for No-Avatar-Found in 'models.Avatar.delete' with bool return.

+++ api/models/user.py
@@ -318,7 +318,9 @@ class Avatar(Base):
         return avatar
 
     @staticmethod
-    async def delete(user_id: str) -> None:
+    async def delete(user_id: str) -> bool:
         avatar = await Avatar.get_by_id(user_id=user_id)
-        await db.delete(avatar)
-        return
+        if avatar:
+            await db.delete(avatar)
+            return True
+        return False
